fixed wrong links when root_dir != DocumentRoot
If the root_dir is != the DocumentRoot then files are not linked directly, but
downloaded instead. This prevents broken links. Trying to guess the
DocumentRoot is not really an option, because due to aliases, rewrites and
fallbacks it is nearly impossible to guess the correct DocumentRoot reliably.

This is related to issue #26