Adjust apache uid and gid
The alpine packages now use the username apache instead of www-data, so
we're adjusting to this.

Also, since we want to use this image with the stock php-apache images,
which uses uid and gid 33 (since it's debian based), we also adjust
things accordingly.

NOTE: At a later point this should be solved with a build or environment
      variable.